+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Block changing node pool instance type from one that supports premium storage to one that does not.
+
 ## [1.6.2] - 2020-10-07
 
 ### Fixed